Refactor SignUp and Login components into one component
